chore: rename browser folder to browserdashboard to help with casing issue
