+++ src/scenes/BrowserDashboard/WebVitals/GaugeArrow.tsx
