+++ src/scenes/BrowserDashboard/WebVitals/WebVitalBarGauge.tsx
