+++ src/scenes/BrowserDashboard/WebVitals/WebVitalGauge.tsx
