+++ src/scenes/BrowserDashboard/WebVitals/WebVitalValue.tsx
