+++ src/scenes/BrowserDashboard/WebVitals/types.ts
