+++ src/scenes/BrowserDashboard/WebVitalsOverTimeRow.tsx
